+++ chats/apps/api/v1/urls.py
@@ -1,9 +1,12 @@
 from django.urls import include, path
 
-from chats.apps.api.v1.rooms.viewsets import RoomsReportViewSet
-from chats.apps.api.v1.dashboard.viewsets import ModelFieldsViewSet, ReportFieldsValidatorViewSet
-from chats.apps.api.v1.routers import router
+from chats.apps.api.v1.dashboard.viewsets import (
+    ModelFieldsViewSet,
+    ReportFieldsValidatorViewSet,
+)
 from chats.apps.api.v1.internal.ai_features.views import FeaturePromptsView
+from chats.apps.api.v1.rooms.viewsets import RoomsReportViewSet
+from chats.apps.api.v1.routers import router
 
 urlpatterns = [
     path("rooms/report/", RoomsReportViewSet.as_view(), name="rooms_report"),
@@ -12,7 +15,15 @@ urlpatterns = [
         FeaturePromptsView.as_view(),
         name="ai_features_prompts",
     ),
-    path("model-fields/", ModelFieldsViewSet.as_view({'get': 'list'}), name="model-fields"),
-    path("custom-report/", ReportFieldsValidatorViewSet.as_view({'post': 'create'}), name="custom-report"),
+    path(
+        "model-fields/",
+        ModelFieldsViewSet.as_view({"get": "list"}),
+        name="model-fields",
+    ),
+    path(
+        "custom-report/",
+        ReportFieldsValidatorViewSet.as_view({"post": "create"}),
+        name="custom-report",
+    ),
     path("", include(router.urls)),
 ]