+++ chats/apps/msgs/usecases/set_msg_external_id.py
@@ -2,7 +2,6 @@ from django.db import transaction
 
 from chats.apps.api.utils import create_reply_index
 from chats.apps.msgs.models import Message, MessageMedia
-from chats.apps.api.utils import create_reply_index
 
 
 class SetMsgExternalIdUseCase:
@@ -25,6 +24,5 @@ class SetMsgExternalIdUseCase:
                     create_reply_index(message)
                 except MessageMedia.DoesNotExist:
                     return
-            except Exception as e:
-                print(f"Error setting external id: {e}")
-                raise e
+            except Exception:
+                return