+++ chats/apps/rooms/tests/test_models.py
@@ -264,3 +264,25 @@ class TestRoomModel(TestCase):
         # added_to_queue_at should change when user is removed
         # as, in practice, the room is added to the queue
         self.assertNotEqual(room.added_to_queue_at, added_to_queue_at)
+
+    def test_add_transfer_to_history(self):
+        user = User.objects.create(email="a@user.com")
+        room = Room.objects.create(queue=self.queue)
+        self.assertEqual(room.full_transfer_history, [])
+        feedback = create_transfer_json(
+            action="transfer",
+            from_=room.queue,
+            to=user,
+        )
+        room.add_transfer_to_history(feedback)
+        self.assertEqual(room.full_transfer_history, [feedback])
+        self.assertEqual(room.transfer_history, feedback)
+
+        other_feedback = create_transfer_json(
+            action="transfer",
+            from_=room.queue,
+            to=user,
+        )
+        room.add_transfer_to_history(other_feedback)
+        self.assertEqual(room.full_transfer_history, [feedback, other_feedback])
+        self.assertEqual(room.transfer_history, other_feedback)