+++ CHANGELOG.md
@@ -1,3 +1,46 @@
+# 3.23.4
+## Add
+  - Adding support for rooms_limit in group sector.
+        
+# 3.23.3
+## Fix
+  - Adding livesum in gauge metric
+
+# 3.23.2
+## Fix
+  - Verify if queues or sectors are already integrated before send to flows.
+  - Changing message status to save in is_read/is_delivered fields
+  - Verify if message index exists in chats before try to get status.
+## Add
+  - New field in sector to save working days
+  - Metrics to create websocket dashboard
+
+# 3.23.1
+## Fix
+  - Rooms waiting time metric
+
+# 3.23.0
+## Add
+  - AI-generated chat rooms history summary
+  - Protocol field in new rooms creation
+
+# 3.22.4
+## Add
+  - Websocket notification to all users in the queue when there is a transfer
+  to another user
+## Fix
+  - Rooms list to do not include the pinned rooms logic when the status filter
+  is used and it is not "ongoing"
+
+# 3.22.3
+## Fix
+  - Online agents query to exclude correctly those who have an active custom status,
+  when not called "In-service"
+
+# 3.22.2
+## Add
+  - Automatic room distribution when closing a custom status
+
 # 3.22.1
 ## Add
   - Queue priority as default for all projects